# schema/openapi.yml
openapi: 3.0.0
info:
  title: PR Compass API
  version: 1.0.0
paths:
  /title-recommend:
    get:
      summary: Get title recommendations
      responses:
        '200':
          description: A list of recommended titles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendResponse'
  /judge:
    post:
      summary: Judge a title
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        '200':
          description: The result of the title judgment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgeResult'
  /similar:
    post:
      summary: Find similar titles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        '200':
          description: A list of similar titles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimilarResponse'
components:
  schemas:
    RecommendResponse:
      type: object
      properties:
        newsTopics:
          type: array
          items:
            type: string
    JudgeResult:
      type: object
      properties:
        newsValue:
          $ref: '#/components/schemas/ValidationResult'
        publicDecency:
          $ref: '#/components/schemas/ValidationResult'
        legalCompliance:
          $ref: '#/components/schemas/ValidationResult'
    ValidationResult:
      type: object
      properties:
        isValid:
          type: boolean
        reason:
          type: string
    SimilarResponse:
      type: object
      properties:
        corporationName:
          type: string
        publishdDatetime:
          type: string
        thumbnailUrl:
          type: string
        postUrl:
          type: string
        title:
          type: string
        like_count:
          type: integer